requirements.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0,<11.0.0
pytest==6.2.5
pytest-asyncio==0.18.1